+++ talpid-core/src/tunnel/wireguard/mod.rs
@@ -16,6 +16,8 @@ mod connectivity_check;
 mod logging;
 mod stats;
 mod wireguard_go;
+#[cfg(target_os = "linux")]
+mod wireguard_kernel;
 
 use self::wireguard_go::WgGoTunnel;
 
@@ -62,12 +64,7 @@ impl WireguardMonitor {
         tun_provider: &mut TunProvider,
         route_manager: &mut routing::RouteManager,
     ) -> Result<WireguardMonitor> {
-        let tunnel = Box::new(WgGoTunnel::start_tunnel(
-            &config,
-            log_path,
-            tun_provider,
-            Self::get_tunnel_routes(config),
-        )?);
+        let tunnel = Self::open_tunnel(&config, log_path, tun_provider, route_manager)?;
         let iface_name = tunnel.get_interface_name().to_string();
         route_manager
             .add_routes(Self::get_routes(&iface_name, &config))
@@ -93,7 +90,7 @@ impl WireguardMonitor {
         let close_sender = monitor.close_msg_sender.clone();
         let mut connectivity_monitor = connectivity_check::ConnectivityMonitor::new(
             gateway,
-            iface_name,
+            iface_name.to_string(),
             Arc::downgrade(&monitor.tunnel),
             pinger_rx,
         )?;
@@ -123,6 +120,34 @@ impl WireguardMonitor {
         });
 
         Ok(monitor)
+    }
+
+    #[cfg_attr(not(target_os = "linux"), allow(unused_variables))]
+    fn open_tunnel(
+        config: &Config,
+        log_path: Option<&Path>,
+        tun_provider: &mut TunProvider,
+        route_manager: &mut routing::RouteManager,
+    ) -> Result<Box<dyn Tunnel>> {
+        #[cfg(target_os = "linux")]
+        match wireguard_kernel::KernelTunnel::new(route_manager.runtime_handle(), config) {
+            Ok(tunnel) => {
+                return Ok(Box::new(tunnel));
+            }
+            Err(err) => {
+                log::error!(
+                    "Failed to setup kernel WireGuard device, falling back to userspace: {}",
+                    err
+                );
+            }
+        };
+
+        Ok(Box::new(WgGoTunnel::start_tunnel(
+            &config,
+            log_path,
+            tun_provider,
+            Self::get_tunnel_routes(config),
+        )?))
     }
 
     /// Returns a close handle for the tunnel
@@ -228,7 +253,7 @@ impl CloseHandle {
 }
 
 pub(crate) trait Tunnel: Send {
-    fn get_interface_name(&self) -> &str;
+    fn get_interface_name(&self) -> String;
     fn stop(self: Box<Self>) -> std::result::Result<(), TunnelError>;
     fn get_tunnel_stats(&self) -> std::result::Result<stats::Stats, TunnelError>;
 }